pyproject.toml
[project]
name = "nomondays-web-app"
version = "0.1.0"
description = "Nomondays Web App - Infrastructure V2.0"
authors = [{name = "VelValerian"}]
dependencies = [
    "django==5.2.12",
    "django-axes",
    "django-debug-toolbar",
    "psycopg",
    "redis",
    "django-redis",
    "pytest-django",
    "pytest-cov",
    "model-bakery",
    "pytest-xdist",
    "pytest-sugar"
]
requires-python = ">=3.13"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--reuse-db --cov=web --cov-report=term-missing --cov-report=html"
pythonpath = ["web"]

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "web/config/*",
    "manage.py"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
